+++ test/playground/pages/config-meta/effect/e2e-test.ts
@@ -1,7 +1,7 @@
 export { testSettingEffect }
 
 import { expect, test } from '@brillout/test-e2e'
-import { retrievePageContext } from '../../../utils/retrievePageContext'
+import { retrievePageContext } from '../retrievePageContext'
 
 function testSettingEffect() {
   test('Setting Effect - Not applied', async () => {
