+++ test/playground/pages/config-meta/effect/serializePageContext.ts
@@ -1,5 +1,5 @@
 import { PageContext } from 'vike/types'
-import { serializePageContext as serializePageContextGeneric } from '../../../utils/serializePageContext'
+import { serializePageContext as serializePageContextGeneric } from '../serializePageContext'
 
 export function serializePageContext(pageContext: PageContext) {
   return serializePageContextGeneric(pageContext, ['settingWithEffect', 'dependentSetting'])
